Refactor package structure for `SSEAgent` and `ChatAgent`, centralize `sseAgent` utility in `ktor` package, and streamline imports across modules.
